Rename MixerProfiler to AudioEngineProfiler
