+++ src/core/CMakeLists.txt
@@ -2,6 +2,7 @@ set(LMMS_SRCS
 	${LMMS_SRCS}
 
 	core/AudioEngine.cpp
+	core/AudioEngineProfiler.cpp
 	core/AutomatableModel.cpp
 	core/AutomationPattern.cpp
 	core/AutomationNode.cpp
@@ -42,7 +43,6 @@ set(LMMS_SRCS
 	core/MeterModel.cpp
 	core/MicroTimer.cpp
 	core/Microtuner.cpp
-	core/MixerProfiler.cpp
 	core/MixerWorkerThread.cpp
 	core/MixHelpers.cpp
 	core/Model.cpp
